Struct google_dialogflow2_beta1::api::GoogleCloudDialogflowV2beta1Intent

pub struct GoogleCloudDialogflowV2beta1Intent {
    pub action: Option<String>,
    pub default_response_platforms: Option<Vec<String>>,
    pub display_name: Option<String>,
    pub end_interaction: Option<bool>,
    pub events: Option<Vec<String>>,
    pub followup_intent_info: Option<Vec<GoogleCloudDialogflowV2beta1IntentFollowupIntentInfo>>,
    pub input_context_names: Option<Vec<String>>,
    pub is_fallback: Option<bool>,
    pub live_agent_handoff: Option<bool>,
    pub messages: Option<Vec<GoogleCloudDialogflowV2beta1IntentMessage>>,
    pub ml_disabled: Option<bool>,
    pub ml_enabled: Option<bool>,
    pub name: Option<String>,
    pub output_contexts: Option<Vec<GoogleCloudDialogflowV2beta1Context>>,
    pub parameters: Option<Vec<GoogleCloudDialogflowV2beta1IntentParameter>>,
    pub parent_followup_intent_name: Option<String>,
    pub priority: Option<i32>,
    pub reset_contexts: Option<bool>,
    pub root_followup_intent_name: Option<String>,
    pub training_phrases: Option<Vec<GoogleCloudDialogflowV2beta1IntentTrainingPhrase>>,
    pub webhook_state: Option<String>,
}

An intent categorizes an end-user's intention for one conversation turn. For each agent, you define many intents, where your combined intents can handle a complete conversation. When an end-user writes or says something, referred to as an end-user expression or end-user input, Dialogflow matches the end-user input to the best intent in your agent. Matching an intent is also known as intent classification. For more information, see the intent guide.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

agent intents create projects (request|response)

agent intents get projects (response)

agent intents patch projects (request|response)

locations agent intents create projects (request|response)

locations agent intents get projects (response)

locations agent intents patch projects (request|response)

Fields
action: Option<String>

Optional. The name of the action associated with the intent. Note: The action name must not contain whitespaces.

default_response_platforms: Option<Vec<String>>

Optional. The list of platforms for which the first responses will be copied from the messages in PLATFORM_UNSPECIFIED (i.e. default platform).

display_name: Option<String>

Required. The name of this intent.

end_interaction: Option<bool>

Optional. Indicates that this intent ends an interaction. Some integrations (e.g., Actions on Google or Dialogflow phone gateway) use this information to close interaction with an end user. Default is false.

events: Option<Vec<String>>

Optional. The collection of event names that trigger the intent. If the collection of input contexts is not empty, all of the contexts must be present in the active user session for an event to trigger this intent. Event names are limited to 150 characters.

followup_intent_info: Option<Vec<GoogleCloudDialogflowV2beta1IntentFollowupIntentInfo>>

Output only. Information about all followup intents that have this intent as a direct or indirect parent. We populate this field only in the output.

input_context_names: Option<Vec<String>>

Optional. The list of context names required for this intent to be triggered. Formats: - projects//agent/sessions/-/contexts/ - projects//locations//agent/sessions/-/contexts/

is_fallback: Option<bool>

Optional. Indicates whether this is a fallback intent.

live_agent_handoff: Option<bool>

Optional. Indicates that a live agent should be brought in to handle the interaction with the user. In most cases, when you set this flag to true, you would also want to set end_interaction to true as well. Default is false.

messages: Option<Vec<GoogleCloudDialogflowV2beta1IntentMessage>>

Optional. The collection of rich messages corresponding to the Response field in the Dialogflow console.

ml_disabled: Option<bool>

Optional. Indicates whether Machine Learning is disabled for the intent. Note: If ml_disabled setting is set to true, then this intent is not taken into account during inference in ML ONLY match mode. Also, auto-markup in the UI is turned off.

ml_enabled: Option<bool>

Optional. Indicates whether Machine Learning is enabled for the intent. Note: If ml_enabled setting is set to false, then this intent is not taken into account during inference in ML ONLY match mode. Also, auto-markup in the UI is turned off. DEPRECATED! Please use ml_disabled field instead. NOTE: If both ml_enabled and ml_disabled are either not set or false, then the default value is determined as follows: - Before April 15th, 2018 the default is: ml_enabled = false / ml_disabled = true. - After April 15th, 2018 the default is: ml_enabled = true / ml_disabled = false.

name: Option<String>

Optional. The unique identifier of this intent. Required for Intents.UpdateIntent and Intents.BatchUpdateIntents methods. Supported formats: - projects//agent/intents/ - projects//locations//agent/intents/

output_contexts: Option<Vec<GoogleCloudDialogflowV2beta1Context>>

Optional. The collection of contexts that are activated when the intent is matched. Context messages in this collection should not set the parameters field. Setting the lifespan_count to 0 will reset the context when the intent is matched. Format: projects//agent/sessions/-/contexts/.

parameters: Option<Vec<GoogleCloudDialogflowV2beta1IntentParameter>>

Optional. The collection of parameters associated with the intent.

parent_followup_intent_name: Option<String>

Optional. The unique identifier of the parent intent in the chain of followup intents. You can set this field when creating an intent, for example with CreateIntent or BatchUpdateIntents, in order to make this intent a followup intent. It identifies the parent followup intent. Format: projects//agent/intents/.

priority: Option<i32>

Optional. The priority of this intent. Higher numbers represent higher priorities. - If the supplied value is unspecified or 0, the service translates the value to 500,000, which corresponds to the Normal priority in the console. - If the supplied value is negative, the intent is ignored in runtime detect intent requests.

reset_contexts: Option<bool>

Optional. Indicates whether to delete all contexts in the current session when this intent is matched.

root_followup_intent_name: Option<String>

Output only. The unique identifier of the root intent in the chain of followup intents. It identifies the correct followup intents chain for this intent. Format: projects//agent/intents/.

training_phrases: Option<Vec<GoogleCloudDialogflowV2beta1IntentTrainingPhrase>>

Optional. The collection of examples that the agent is trained on.

webhook_state: Option<String>

Optional. Indicates whether webhooks are enabled for the intent.