Struct google_dialogflow2_beta1::api::GoogleCloudDialogflowV2beta1QueryResult

pub struct GoogleCloudDialogflowV2beta1QueryResult {
    pub action: Option<String>,
    pub all_required_params_present: Option<bool>,
    pub diagnostic_info: Option<HashMap<String, String>>,
    pub fulfillment_messages: Option<Vec<GoogleCloudDialogflowV2beta1IntentMessage>>,
    pub fulfillment_text: Option<String>,
    pub intent: Option<GoogleCloudDialogflowV2beta1Intent>,
    pub intent_detection_confidence: Option<f32>,
    pub knowledge_answers: Option<GoogleCloudDialogflowV2beta1KnowledgeAnswers>,
    pub language_code: Option<String>,
    pub output_contexts: Option<Vec<GoogleCloudDialogflowV2beta1Context>>,
    pub parameters: Option<HashMap<String, String>>,
    pub query_text: Option<String>,
    pub sentiment_analysis_result: Option<GoogleCloudDialogflowV2beta1SentimentAnalysisResult>,
    pub speech_recognition_confidence: Option<f32>,
    pub webhook_payload: Option<HashMap<String, String>>,
    pub webhook_source: Option<String>,
}

Represents the result of conversational query or event processing.

This type is not used in any activity, and only used as part of another schema.

Fields
action: Option<String>

The action name from the matched intent.

all_required_params_present: Option<bool>

This field is set to: - false if the matched intent has required parameters and not all of the required parameter values have been collected. - true if all required parameter values have been collected, or if the matched intent doesn't contain any required parameters.

diagnostic_info: Option<HashMap<String, String>>

Free-form diagnostic information for the associated detect intent request. The fields of this data can change without notice, so you should not write code that depends on its structure. The data may contain: - webhook call latency - webhook errors

fulfillment_messages: Option<Vec<GoogleCloudDialogflowV2beta1IntentMessage>>

The collection of rich messages to present to the user.

fulfillment_text: Option<String>

The text to be pronounced to the user or shown on the screen. Note: This is a legacy field, fulfillment_messages should be preferred.

intent: Option<GoogleCloudDialogflowV2beta1Intent>

The intent that matched the conversational query. Some, not all fields are filled in this message, including but not limited to: name, display_name, end_interaction and is_fallback.

intent_detection_confidence: Option<f32>

The intent detection confidence. Values range from 0.0 (completely uncertain) to 1.0 (completely certain). This value is for informational purpose only and is only used to help match the best intent within the classification threshold. This value may change for the same end-user expression at any time due to a model retraining or change in implementation. If there are multiple knowledge_answers messages, this value is set to the greatest knowledgeAnswers.match_confidence value in the list.

knowledge_answers: Option<GoogleCloudDialogflowV2beta1KnowledgeAnswers>

The result from Knowledge Connector (if any), ordered by decreasing KnowledgeAnswers.match_confidence.

language_code: Option<String>

The language that was triggered during intent detection. See Language Support for a list of the currently supported language codes.

output_contexts: Option<Vec<GoogleCloudDialogflowV2beta1Context>>

The collection of output contexts. If applicable, output_contexts.parameters contains entries with name .original containing the original parameter values before the query.

parameters: Option<HashMap<String, String>>

The collection of extracted parameters. Depending on your protocol or client library language, this is a map, associative array, symbol table, dictionary, or JSON object composed of a collection of (MapKey, MapValue) pairs: - MapKey type: string - MapKey value: parameter name - MapValue type: - If parameter's entity type is a composite entity: map - Else: depending on parameter value type, could be one of string, number, boolean, null, list or map - MapValue value: - If parameter's entity type is a composite entity: map from composite entity property names to property values - Else: parameter value

query_text: Option<String>

The original conversational query text: - If natural language text was provided as input, query_text contains a copy of the input. - If natural language speech audio was provided as input, query_text contains the speech recognition result. If speech recognizer produced multiple alternatives, a particular one is picked. - If automatic spell correction is enabled, query_text will contain the corrected user input.

sentiment_analysis_result: Option<GoogleCloudDialogflowV2beta1SentimentAnalysisResult>

The sentiment analysis result, which depends on the sentiment_analysis_request_config specified in the request.

speech_recognition_confidence: Option<f32>

The Speech recognition confidence between 0.0 and 1.0. A higher number indicates an estimated greater likelihood that the recognized words are correct. The default of 0.0 is a sentinel value indicating that confidence was not set. This field is not guaranteed to be accurate or set. In particular this field isn't set for StreamingDetectIntent since the streaming endpoint has separate confidence estimates per portion of the audio in StreamingRecognitionResult.

webhook_payload: Option<HashMap<String, String>>

If the query was fulfilled by a webhook call, this field is set to the value of the payload field returned in the webhook response.

webhook_source: Option<String>

If the query was fulfilled by a webhook call, this field is set to the value of the source field returned in the webhook response.