Struct google_dialogflow2_beta1::ProjectMethods

pub struct ProjectMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the Dialogflow hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_dialogflow2_beta1 as dialogflow2_beta1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dialogflow2_beta1::Dialogflow;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Dialogflow::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `agent(...)`, `agent_entity_types_batch_delete(...)`, `agent_entity_types_batch_update(...)`, `agent_entity_types_create(...)`, `agent_entity_types_delete(...)`, `agent_entity_types_entities_batch_create(...)`, `agent_entity_types_entities_batch_delete(...)`, `agent_entity_types_entities_batch_update(...)`, `agent_entity_types_get(...)`, `agent_entity_types_list(...)`, `agent_entity_types_patch(...)`, `agent_environments_users_sessions_contexts_create(...)`, `agent_environments_users_sessions_contexts_delete(...)`, `agent_environments_users_sessions_contexts_get(...)`, `agent_environments_users_sessions_contexts_list(...)`, `agent_environments_users_sessions_contexts_patch(...)`, `agent_environments_users_sessions_delete_contexts(...)`, `agent_environments_users_sessions_detect_intent(...)`, `agent_environments_users_sessions_entity_types_create(...)`, `agent_environments_users_sessions_entity_types_delete(...)`, `agent_environments_users_sessions_entity_types_get(...)`, `agent_environments_users_sessions_entity_types_list(...)`, `agent_environments_users_sessions_entity_types_patch(...)`, `agent_export(...)`, `agent_import(...)`, `agent_intents_batch_delete(...)`, `agent_intents_batch_update(...)`, `agent_intents_create(...)`, `agent_intents_delete(...)`, `agent_intents_get(...)`, `agent_intents_list(...)`, `agent_intents_patch(...)`, `agent_knowledge_bases_create(...)`, `agent_knowledge_bases_delete(...)`, `agent_knowledge_bases_documents_create(...)`, `agent_knowledge_bases_documents_delete(...)`, `agent_knowledge_bases_documents_get(...)`, `agent_knowledge_bases_documents_list(...)`, `agent_knowledge_bases_documents_patch(...)`, `agent_knowledge_bases_documents_reload(...)`, `agent_knowledge_bases_get(...)`, `agent_knowledge_bases_list(...)`, `agent_knowledge_bases_patch(...)`, `agent_restore(...)`, `agent_search(...)`, `agent_sessions_contexts_create(...)`, `agent_sessions_contexts_delete(...)`, `agent_sessions_contexts_get(...)`, `agent_sessions_contexts_list(...)`, `agent_sessions_contexts_patch(...)`, `agent_sessions_delete_contexts(...)`, `agent_sessions_detect_intent(...)`, `agent_sessions_entity_types_create(...)`, `agent_sessions_entity_types_delete(...)`, `agent_sessions_entity_types_get(...)`, `agent_sessions_entity_types_list(...)`, `agent_sessions_entity_types_patch(...)`, `agent_train(...)`, `delete_agent(...)`, `get_agent(...)`, `knowledge_bases_create(...)`, `knowledge_bases_delete(...)`, `knowledge_bases_documents_create(...)`, `knowledge_bases_documents_delete(...)`, `knowledge_bases_documents_get(...)`, `knowledge_bases_documents_list(...)`, `knowledge_bases_documents_patch(...)`, `knowledge_bases_documents_reload(...)`, `knowledge_bases_get(...)`, `knowledge_bases_list(...)`, `knowledge_bases_patch(...)` and `operations_get(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

pub fn knowledge_bases_list(
&self,
parent: &str
) -> ProjectKnowledgeBaseListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the list of all knowledge bases of the specified agent.

Note: The projects.agent.knowledgeBases resource is deprecated; only use projects.knowledgeBases.

Arguments

parent - Required. The project to list of knowledge bases for. Format: projects/<Project ID>.

pub fn agent_environments_users_sessions_contexts_delete(
&self,
name: &str
) -> ProjectAgentEnvironmentUserSessionContextDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified context.

Arguments

name - Required. The name of the context to delete. Format: projects/<Project ID>/agent/sessions/<Session ID>/contexts/<Context ID> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>/contexts/<Context ID>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn agent_environments_users_sessions_entity_types_delete(
&self,
name: &str
) -> ProjectAgentEnvironmentUserSessionEntityTypeDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified session entity type.

Arguments

name - Required. The name of the entity type to delete. Format: projects/<Project ID>/agent/sessions/<Session ID>/entityTypes/<Entity Type Display Name> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>/entityTypes/<Entity Type Display Name>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn agent_knowledge_bases_patch(
&self,
request: GoogleCloudDialogflowV2beta1KnowledgeBase,
name: &str
) -> ProjectAgentKnowledgeBasePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified knowledge base.

Note: The projects.agent.knowledgeBases resource is deprecated; only use projects.knowledgeBases.

Arguments

request - No description provided.

name - The knowledge base resource name. The name must be empty when creating a knowledge base. Format: projects/<Project ID>/knowledgeBases/<Knowledge Base ID>.

pub fn agent_entity_types_create(
&self,
request: GoogleCloudDialogflowV2beta1EntityType,
parent: &str
) -> ProjectAgentEntityTypeCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an entity type in the specified agent.

Arguments

request - No description provided.

parent - Required. The agent to create a entity type for. Format: projects/<Project ID>/agent.

pub fn agent_environments_users_sessions_contexts_get(
&self,
name: &str
) -> ProjectAgentEnvironmentUserSessionContextGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the specified context.

Arguments

name - Required. The name of the context. Format: projects/<Project ID>/agent/sessions/<Session ID>/contexts/<Context ID> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>/contexts/<Context ID>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn agent_knowledge_bases_delete(
&self,
name: &str
) -> ProjectAgentKnowledgeBaseDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified knowledge base.

Note: The projects.agent.knowledgeBases resource is deprecated; only use projects.knowledgeBases.

Arguments

name - Required. The name of the knowledge base to delete. Format: projects/<Project ID>/knowledgeBases/<Knowledge Base ID>.

pub fn agent_knowledge_bases_documents_delete(
&self,
name: &str
) -> ProjectAgentKnowledgeBaseDocumentDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified document.

Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents.

Operation <response: google.protobuf.Empty, metadata: KnowledgeOperationMetadata>

Arguments

name - The name of the document to delete. Format: projects/<Project ID>/knowledgeBases/<Knowledge Base ID>/documents/<Document ID>.

pub fn agent_sessions_entity_types_patch(
&self,
request: GoogleCloudDialogflowV2beta1SessionEntityType,
name: &str
) -> ProjectAgentSessionEntityTypePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified session entity type.

Arguments

request - No description provided.

name - Required. The unique identifier of this session entity type. Format: projects/<Project ID>/agent/sessions/<Session ID>/entityTypes/<Entity Type Display Name>, or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>/entityTypes/<Entity Type Display Name>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user. <Entity Type Display Name> must be the display name of an existing entity type in the same agent that will be overridden or supplemented.

pub fn agent_entity_types_batch_update(
&self,
request: GoogleCloudDialogflowV2beta1BatchUpdateEntityTypesRequest,
parent: &str
) -> ProjectAgentEntityTypeBatchUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates/Creates multiple entity types in the specified agent.

Operation <response: BatchUpdateEntityTypesResponse>

Arguments

request - No description provided.

parent - Required. The name of the agent to update or create entity types in. Format: projects/<Project ID>/agent.

pub fn agent_knowledge_bases_create(
&self,
request: GoogleCloudDialogflowV2beta1KnowledgeBase,
parent: &str
) -> ProjectAgentKnowledgeBaseCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a knowledge base.

Note: The projects.agent.knowledgeBases resource is deprecated; only use projects.knowledgeBases.

Arguments

request - No description provided.

parent - Required. The project to create a knowledge base for. Format: projects/<Project ID>.

pub fn agent_import(
&self,
request: GoogleCloudDialogflowV2beta1ImportAgentRequest,
parent: &str
) -> ProjectAgentImportCall<'a, C, A>

Create a builder to help you perform the following task:

Imports the specified agent from a ZIP file.

Uploads new intents and entity types without deleting the existing ones. Intents and entity types with the same name are replaced with the new versions from ImportAgentRequest.

Operation <response: google.protobuf.Empty>

Arguments

request - No description provided.

parent - Required. The project that the agent to import is associated with. Format: projects/<Project ID>.

pub fn agent_intents_batch_update(
&self,
request: GoogleCloudDialogflowV2beta1BatchUpdateIntentsRequest,
parent: &str
) -> ProjectAgentIntentBatchUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates/Creates multiple intents in the specified agent.

Operation <response: BatchUpdateIntentsResponse>

Arguments

request - No description provided.

parent - Required. The name of the agent to update or create intents in. Format: projects/<Project ID>/agent.

pub fn agent_intents_batch_delete(
&self,
request: GoogleCloudDialogflowV2beta1BatchDeleteIntentsRequest,
parent: &str
) -> ProjectAgentIntentBatchDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes intents in the specified agent.

Operation <response: google.protobuf.Empty>

Arguments

request - No description provided.

parent - Required. The name of the agent to delete all entities types for. Format: projects/<Project ID>/agent.

pub fn agent_knowledge_bases_get(
&self,
name: &str
) -> ProjectAgentKnowledgeBaseGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the specified knowledge base.

Note: The projects.agent.knowledgeBases resource is deprecated; only use projects.knowledgeBases.

Arguments

name - Required. The name of the knowledge base to retrieve. Format projects/<Project ID>/knowledgeBases/<Knowledge Base ID>.

pub fn agent_restore(
&self,
request: GoogleCloudDialogflowV2beta1RestoreAgentRequest,
parent: &str
) -> ProjectAgentRestoreCall<'a, C, A>

Create a builder to help you perform the following task:

Restores the specified agent from a ZIP file.

Replaces the current agent version with a new one. All the intents and entity types in the older version are deleted.

Operation <response: google.protobuf.Empty>

Arguments

request - No description provided.

parent - Required. The project that the agent to restore is associated with. Format: projects/<Project ID>.

pub fn agent_knowledge_bases_documents_patch(
&self,
request: GoogleCloudDialogflowV2beta1Document,
name: &str
) -> ProjectAgentKnowledgeBaseDocumentPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified document.

Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents.

Operation <response: Document, metadata: KnowledgeOperationMetadata>

Arguments

request - No description provided.

name - The document resource name. The name must be empty when creating a document. Format: projects/<Project ID>/knowledgeBases/<Knowledge Base ID>/documents/<Document ID>.

pub fn agent_sessions_detect_intent(
&self,
request: GoogleCloudDialogflowV2beta1DetectIntentRequest,
session: &str
) -> ProjectAgentSessionDetectIntentCall<'a, C, A>

Create a builder to help you perform the following task:

Processes a natural language query and returns structured, actionable data as a result. This method is not idempotent, because it may cause contexts and session entity types to be updated, which in turn might affect results of future queries.

Arguments

request - No description provided.

session - Required. The name of the session this query is sent to. Format: projects/<Project ID>/agent/sessions/<Session ID>, or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we are using "-". It's up to the API caller to choose an appropriate Session ID and User Id. They can be a random number or some type of user and session identifiers (preferably hashed). The length of the Session ID and User ID must not exceed 36 characters.

pub fn knowledge_bases_documents_create(
&self,
request: GoogleCloudDialogflowV2beta1Document,
parent: &str
) -> ProjectKnowledgeBaseDocumentCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new document.

Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents.

Operation <response: Document, metadata: KnowledgeOperationMetadata>

Arguments

request - No description provided.

parent - Required. The knoweldge base to create a document for. Format: projects/<Project ID>/knowledgeBases/<Knowledge Base ID>.

pub fn agent_sessions_contexts_delete(
&self,
name: &str
) -> ProjectAgentSessionContextDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified context.

Arguments

name - Required. The name of the context to delete. Format: projects/<Project ID>/agent/sessions/<Session ID>/contexts/<Context ID> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>/contexts/<Context ID>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn agent_environments_users_sessions_contexts_patch(
&self,
request: GoogleCloudDialogflowV2beta1Context,
name: &str
) -> ProjectAgentEnvironmentUserSessionContextPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified context.

Arguments

request - No description provided.

name - Required. The unique identifier of the context. Format: projects/<Project ID>/agent/sessions/<Session ID>/contexts/<Context ID>, or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>/contexts/<Context ID>. The Context ID is always converted to lowercase, may only contain characters in a-zA-Z0-9_-% and may be at most 250 bytes long. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn get_agent(&self, parent: &str) -> ProjectGetAgentCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the specified agent.

Arguments

parent - Required. The project that the agent to fetch is associated with. Format: projects/<Project ID>.

pub fn agent_environments_users_sessions_entity_types_patch(
&self,
request: GoogleCloudDialogflowV2beta1SessionEntityType,
name: &str
) -> ProjectAgentEnvironmentUserSessionEntityTypePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified session entity type.

Arguments

request - No description provided.

name - Required. The unique identifier of this session entity type. Format: projects/<Project ID>/agent/sessions/<Session ID>/entityTypes/<Entity Type Display Name>, or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>/entityTypes/<Entity Type Display Name>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user. <Entity Type Display Name> must be the display name of an existing entity type in the same agent that will be overridden or supplemented.

pub fn agent(
&self,
request: GoogleCloudDialogflowV2beta1Agent,
parent: &str
) -> ProjectAgentCall<'a, C, A>

Create a builder to help you perform the following task:

Creates/updates the specified agent.

Arguments

request - No description provided.

parent - Required. The project of this agent. Format: projects/<Project ID>.

pub fn agent_entity_types_batch_delete(
&self,
request: GoogleCloudDialogflowV2beta1BatchDeleteEntityTypesRequest,
parent: &str
) -> ProjectAgentEntityTypeBatchDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes entity types in the specified agent.

Operation <response: google.protobuf.Empty>

Arguments

request - No description provided.

parent - Required. The name of the agent to delete all entities types for. Format: projects/<Project ID>/agent.

pub fn knowledge_bases_delete(
&self,
name: &str
) -> ProjectKnowledgeBaseDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified knowledge base.

Note: The projects.agent.knowledgeBases resource is deprecated; only use projects.knowledgeBases.

Arguments

name - Required. The name of the knowledge base to delete. Format: projects/<Project ID>/knowledgeBases/<Knowledge Base ID>.

pub fn agent_knowledge_bases_list(
&self,
parent: &str
) -> ProjectAgentKnowledgeBaseListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the list of all knowledge bases of the specified agent.

Note: The projects.agent.knowledgeBases resource is deprecated; only use projects.knowledgeBases.

Arguments

parent - Required. The project to list of knowledge bases for. Format: projects/<Project ID>.

pub fn agent_export(
&self,
request: GoogleCloudDialogflowV2beta1ExportAgentRequest,
parent: &str
) -> ProjectAgentExportCall<'a, C, A>

Create a builder to help you perform the following task:

Exports the specified agent to a ZIP file.

Operation <response: ExportAgentResponse>

Arguments

request - No description provided.

parent - Required. The project that the agent to export is associated with. Format: projects/<Project ID>.

pub fn agent_knowledge_bases_documents_reload(
&self,
request: GoogleCloudDialogflowV2beta1ReloadDocumentRequest,
name: &str
) -> ProjectAgentKnowledgeBaseDocumentReloadCall<'a, C, A>

Create a builder to help you perform the following task:

Reloads the specified document from its specified source, content_uri or content. The previously loaded content of the document will be deleted. Note: Even when the content of the document has not changed, there still may be side effects because of internal implementation changes.

Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents.

Operation <response: Document, metadata: KnowledgeOperationMetadata>

Arguments

request - No description provided.

name - The name of the document to reload. Format: projects/<Project ID>/knowledgeBases/<Knowledge Base ID>/documents/<Document ID>

pub fn agent_intents_list(
&self,
parent: &str
) -> ProjectAgentIntentListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the list of all intents in the specified agent.

Arguments

parent - Required. The agent to list all intents from. Format: projects/<Project ID>/agent.

pub fn agent_entity_types_entities_batch_update(
&self,
request: GoogleCloudDialogflowV2beta1BatchUpdateEntitiesRequest,
parent: &str
) -> ProjectAgentEntityTypeEntityBatchUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates or creates multiple entities in the specified entity type. This method does not affect entities in the entity type that aren't explicitly specified in the request.

Operation <response: google.protobuf.Empty>

Arguments

request - No description provided.

parent - Required. The name of the entity type to update or create entities in. Format: projects/<Project ID>/agent/entityTypes/<Entity Type ID>.

pub fn agent_sessions_entity_types_create(
&self,
request: GoogleCloudDialogflowV2beta1SessionEntityType,
parent: &str
) -> ProjectAgentSessionEntityTypeCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a session entity type.

If the specified session entity type already exists, overrides the session entity type.

Arguments

request - No description provided.

parent - Required. The session to create a session entity type for. Format: projects/<Project ID>/agent/sessions/<Session ID> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/ sessions/<Session ID>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn operations_get(&self, name: &str) -> ProjectOperationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn agent_environments_users_sessions_delete_contexts(
&self,
parent: &str
) -> ProjectAgentEnvironmentUserSessionDeleteContextCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes all active contexts in the specified session.

Arguments

parent - Required. The name of the session to delete all contexts from. Format: projects/<Project ID>/agent/sessions/<Session ID> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>. If Environment ID is not specified we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn agent_entity_types_entities_batch_create(
&self,
request: GoogleCloudDialogflowV2beta1BatchCreateEntitiesRequest,
parent: &str
) -> ProjectAgentEntityTypeEntityBatchCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates multiple new entities in the specified entity type.

Operation <response: google.protobuf.Empty>

Arguments

request - No description provided.

parent - Required. The name of the entity type to create entities in. Format: projects/<Project ID>/agent/entityTypes/<Entity Type ID>.

pub fn agent_entity_types_list(
&self,
parent: &str
) -> ProjectAgentEntityTypeListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the list of all entity types in the specified agent.

Arguments

parent - Required. The agent to list all entity types from. Format: projects/<Project ID>/agent.

pub fn agent_knowledge_bases_documents_get(
&self,
name: &str
) -> ProjectAgentKnowledgeBaseDocumentGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the specified document.

Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents.

Arguments

name - Required. The name of the document to retrieve. Format projects/<Project ID>/knowledgeBases/<Knowledge Base ID>/documents/<Document ID>.

pub fn knowledge_bases_get(
&self,
name: &str
) -> ProjectKnowledgeBaseGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the specified knowledge base.

Note: The projects.agent.knowledgeBases resource is deprecated; only use projects.knowledgeBases.

Arguments

name - Required. The name of the knowledge base to retrieve. Format projects/<Project ID>/knowledgeBases/<Knowledge Base ID>.

pub fn agent_environments_users_sessions_entity_types_create(
&self,
request: GoogleCloudDialogflowV2beta1SessionEntityType,
parent: &str
) -> ProjectAgentEnvironmentUserSessionEntityTypeCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a session entity type.

If the specified session entity type already exists, overrides the session entity type.

Arguments

request - No description provided.

parent - Required. The session to create a session entity type for. Format: projects/<Project ID>/agent/sessions/<Session ID> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/ sessions/<Session ID>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn agent_environments_users_sessions_entity_types_list(
&self,
parent: &str
) -> ProjectAgentEnvironmentUserSessionEntityTypeListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the list of all session entity types in the specified session.

Arguments

parent - Required. The session to list all session entity types from. Format: projects/<Project ID>/agent/sessions/<Session ID> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/ sessions/<Session ID>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn agent_environments_users_sessions_detect_intent(
&self,
request: GoogleCloudDialogflowV2beta1DetectIntentRequest,
session: &str
) -> ProjectAgentEnvironmentUserSessionDetectIntentCall<'a, C, A>

Create a builder to help you perform the following task:

Processes a natural language query and returns structured, actionable data as a result. This method is not idempotent, because it may cause contexts and session entity types to be updated, which in turn might affect results of future queries.

Arguments

request - No description provided.

session - Required. The name of the session this query is sent to. Format: projects/<Project ID>/agent/sessions/<Session ID>, or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we are using "-". It's up to the API caller to choose an appropriate Session ID and User Id. They can be a random number or some type of user and session identifiers (preferably hashed). The length of the Session ID and User ID must not exceed 36 characters.

pub fn agent_environments_users_sessions_contexts_create(
&self,
request: GoogleCloudDialogflowV2beta1Context,
parent: &str
) -> ProjectAgentEnvironmentUserSessionContextCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a context.

If the specified context already exists, overrides the context.

Arguments

request - No description provided.

parent - Required. The session to create a context for. Format: projects/<Project ID>/agent/sessions/<Session ID> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn agent_entity_types_patch(
&self,
request: GoogleCloudDialogflowV2beta1EntityType,
name: &str
) -> ProjectAgentEntityTypePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified entity type.

Arguments

request - No description provided.

name - The unique identifier of the entity type. Required for EntityTypes.UpdateEntityType and EntityTypes.BatchUpdateEntityTypes methods. Format: projects/<Project ID>/agent/entityTypes/<Entity Type ID>.

pub fn agent_environments_users_sessions_contexts_list(
&self,
parent: &str
) -> ProjectAgentEnvironmentUserSessionContextListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the list of all contexts in the specified session.

Arguments

parent - Required. The session to list all contexts from. Format: projects/<Project ID>/agent/sessions/<Session ID> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn agent_entity_types_delete(
&self,
name: &str
) -> ProjectAgentEntityTypeDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified entity type.

Arguments

name - Required. The name of the entity type to delete. Format: projects/<Project ID>/agent/entityTypes/<EntityType ID>.

pub fn agent_sessions_entity_types_get(
&self,
name: &str
) -> ProjectAgentSessionEntityTypeGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the specified session entity type.

Arguments

name - Required. The name of the session entity type. Format: projects/<Project ID>/agent/sessions/<Session ID>/entityTypes/<Entity Type Display Name> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>/entityTypes/<Entity Type Display Name>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn knowledge_bases_documents_get(
&self,
name: &str
) -> ProjectKnowledgeBaseDocumentGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the specified document.

Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents.

Arguments

name - Required. The name of the document to retrieve. Format projects/<Project ID>/knowledgeBases/<Knowledge Base ID>/documents/<Document ID>.

pub fn agent_intents_get(
&self,
name: &str
) -> ProjectAgentIntentGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the specified intent.

Arguments

name - Required. The name of the intent. Format: projects/<Project ID>/agent/intents/<Intent ID>.

pub fn agent_entity_types_get(
&self,
name: &str
) -> ProjectAgentEntityTypeGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the specified entity type.

Arguments

name - Required. The name of the entity type. Format: projects/<Project ID>/agent/entityTypes/<EntityType ID>.

pub fn agent_entity_types_entities_batch_delete(
&self,
request: GoogleCloudDialogflowV2beta1BatchDeleteEntitiesRequest,
parent: &str
) -> ProjectAgentEntityTypeEntityBatchDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes entities in the specified entity type.

Operation <response: google.protobuf.Empty>

Arguments

request - No description provided.

parent - Required. The name of the entity type to delete entries for. Format: projects/<Project ID>/agent/entityTypes/<Entity Type ID>.

pub fn agent_sessions_contexts_patch(
&self,
request: GoogleCloudDialogflowV2beta1Context,
name: &str
) -> ProjectAgentSessionContextPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified context.

Arguments

request - No description provided.

name - Required. The unique identifier of the context. Format: projects/<Project ID>/agent/sessions/<Session ID>/contexts/<Context ID>, or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>/contexts/<Context ID>. The Context ID is always converted to lowercase, may only contain characters in a-zA-Z0-9_-% and may be at most 250 bytes long. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn agent_knowledge_bases_documents_create(
&self,
request: GoogleCloudDialogflowV2beta1Document,
parent: &str
) -> ProjectAgentKnowledgeBaseDocumentCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new document.

Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents.

Operation <response: Document, metadata: KnowledgeOperationMetadata>

Arguments

request - No description provided.

parent - Required. The knoweldge base to create a document for. Format: projects/<Project ID>/knowledgeBases/<Knowledge Base ID>.

pub fn agent_sessions_contexts_get(
&self,
name: &str
) -> ProjectAgentSessionContextGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the specified context.

Arguments

name - Required. The name of the context. Format: projects/<Project ID>/agent/sessions/<Session ID>/contexts/<Context ID> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>/contexts/<Context ID>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn agent_sessions_delete_contexts(
&self,
parent: &str
) -> ProjectAgentSessionDeleteContextCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes all active contexts in the specified session.

Arguments

parent - Required. The name of the session to delete all contexts from. Format: projects/<Project ID>/agent/sessions/<Session ID> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>. If Environment ID is not specified we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn knowledge_bases_documents_patch(
&self,
request: GoogleCloudDialogflowV2beta1Document,
name: &str
) -> ProjectKnowledgeBaseDocumentPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified document.

Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents.

Operation <response: Document, metadata: KnowledgeOperationMetadata>

Arguments

request - No description provided.

name - The document resource name. The name must be empty when creating a document. Format: projects/<Project ID>/knowledgeBases/<Knowledge Base ID>/documents/<Document ID>.

pub fn knowledge_bases_documents_list(
&self,
parent: &str
) -> ProjectKnowledgeBaseDocumentListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the list of all documents of the knowledge base.

Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents.

Arguments

parent - Required. The knowledge base to list all documents for. Format: projects/<Project ID>/knowledgeBases/<Knowledge Base ID>.

pub fn knowledge_bases_documents_delete(
&self,
name: &str
) -> ProjectKnowledgeBaseDocumentDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified document.

Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents.

Operation <response: google.protobuf.Empty, metadata: KnowledgeOperationMetadata>

Arguments

name - The name of the document to delete. Format: projects/<Project ID>/knowledgeBases/<Knowledge Base ID>/documents/<Document ID>.

pub fn agent_sessions_contexts_create(
&self,
request: GoogleCloudDialogflowV2beta1Context,
parent: &str
) -> ProjectAgentSessionContextCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a context.

If the specified context already exists, overrides the context.

Arguments

request - No description provided.

parent - Required. The session to create a context for. Format: projects/<Project ID>/agent/sessions/<Session ID> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn delete_agent(&self, parent: &str) -> ProjectDeleteAgentCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified agent.

Arguments

parent - Required. The project that the agent to delete is associated with. Format: projects/<Project ID>.

pub fn agent_train(
&self,
request: GoogleCloudDialogflowV2beta1TrainAgentRequest,
parent: &str
) -> ProjectAgentTrainCall<'a, C, A>

Create a builder to help you perform the following task:

Trains the specified agent.

Operation <response: google.protobuf.Empty>

Arguments

request - No description provided.

parent - Required. The project that the agent to train is associated with. Format: projects/<Project ID>.

pub fn agent_sessions_entity_types_list(
&self,
parent: &str
) -> ProjectAgentSessionEntityTypeListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the list of all session entity types in the specified session.

Arguments

parent - Required. The session to list all session entity types from. Format: projects/<Project ID>/agent/sessions/<Session ID> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/ sessions/<Session ID>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn agent_environments_users_sessions_entity_types_get(
&self,
name: &str
) -> ProjectAgentEnvironmentUserSessionEntityTypeGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the specified session entity type.

Arguments

name - Required. The name of the session entity type. Format: projects/<Project ID>/agent/sessions/<Session ID>/entityTypes/<Entity Type Display Name> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>/entityTypes/<Entity Type Display Name>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn knowledge_bases_patch(
&self,
request: GoogleCloudDialogflowV2beta1KnowledgeBase,
name: &str
) -> ProjectKnowledgeBasePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified knowledge base.

Note: The projects.agent.knowledgeBases resource is deprecated; only use projects.knowledgeBases.

Arguments

request - No description provided.

name - The knowledge base resource name. The name must be empty when creating a knowledge base. Format: projects/<Project ID>/knowledgeBases/<Knowledge Base ID>.

pub fn knowledge_bases_create(
&self,
request: GoogleCloudDialogflowV2beta1KnowledgeBase,
parent: &str
) -> ProjectKnowledgeBaseCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a knowledge base.

Note: The projects.agent.knowledgeBases resource is deprecated; only use projects.knowledgeBases.

Arguments

request - No description provided.

parent - Required. The project to create a knowledge base for. Format: projects/<Project ID>.

pub fn agent_knowledge_bases_documents_list(
&self,
parent: &str
) -> ProjectAgentKnowledgeBaseDocumentListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the list of all documents of the knowledge base.

Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents.

Arguments

parent - Required. The knowledge base to list all documents for. Format: projects/<Project ID>/knowledgeBases/<Knowledge Base ID>.

pub fn agent_intents_patch(
&self,
request: GoogleCloudDialogflowV2beta1Intent,
name: &str
) -> ProjectAgentIntentPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified intent.

Arguments

request - No description provided.

name - The unique identifier of this intent. Required for Intents.UpdateIntent and Intents.BatchUpdateIntents methods. Format: projects/<Project ID>/agent/intents/<Intent ID>.

pub fn agent_sessions_entity_types_delete(
&self,
name: &str
) -> ProjectAgentSessionEntityTypeDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified session entity type.

Arguments

name - Required. The name of the entity type to delete. Format: projects/<Project ID>/agent/sessions/<Session ID>/entityTypes/<Entity Type Display Name> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>/entityTypes/<Entity Type Display Name>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.

pub fn agent_search(&self, parent: &str) -> ProjectAgentSearchCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the list of agents.

Since there is at most one conversational agent per project, this method is useful primarily for listing all agents across projects the caller has access to. One can achieve that with a wildcard project collection id "-". Refer to List Sub-Collections.

Arguments

parent - Required. The project to list agents from. Format: projects/<Project ID or '-'>.

pub fn agent_intents_delete(
&self,
name: &str
) -> ProjectAgentIntentDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified intent and its direct or indirect followup intents.

Arguments

name - Required. The name of the intent to delete. If this intent has direct or indirect followup intents, we also delete them. Format: projects/<Project ID>/agent/intents/<Intent ID>.

pub fn knowledge_bases_documents_reload(
&self,
request: GoogleCloudDialogflowV2beta1ReloadDocumentRequest,
name: &str
) -> ProjectKnowledgeBaseDocumentReloadCall<'a, C, A>

Create a builder to help you perform the following task:

Reloads the specified document from its specified source, content_uri or content. The previously loaded content of the document will be deleted. Note: Even when the content of the document has not changed, there still may be side effects because of internal implementation changes.

Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents.

Operation <response: Document, metadata: KnowledgeOperationMetadata>

Arguments

request - No description provided.

name - The name of the document to reload. Format: projects/<Project ID>/knowledgeBases/<Knowledge Base ID>/documents/<Document ID>

pub fn agent_intents_create(
&self,
request: GoogleCloudDialogflowV2beta1Intent,
parent: &str
) -> ProjectAgentIntentCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an intent in the specified agent.

Arguments

request - No description provided.

parent - Required. The agent to create a intent for. Format: projects/<Project ID>/agent.

pub fn agent_sessions_contexts_list(
&self,
parent: &str
) -> ProjectAgentSessionContextListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the list of all contexts in the specified session.

Arguments

parent - Required. The session to list all contexts from. Format: projects/<Project ID>/agent/sessions/<Session ID> or projects/<Project ID>/agent/environments/<Environment ID>/users/<User ID>/sessions/<Session ID>. If Environment ID is not specified, we assume default 'draft' environment. If User ID is not specified, we assume default '-' user.